pyproject.toml
[project]
name = "annex4ac"
version = "1.1.0"
description = "Annex IV-as-Code CLI: generate & validate EU AI Act Annex IV with legal compliance"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
license-files = ["LICENSE", "THIRD_PARTY_LICENSES.md", "annex4ac/fonts/OFL.txt"]
authors = [{name = "Aleksandr Racionaluss", email = "prihodko02bk@gmail.com"}]
keywords = ["AI Act", "compliance", "CLI"]
dependencies = [
  "typer[all]>=0.12",
  "pydantic>=2.7",
  "requests>=2.32",
  "ruamel.yaml>=0.18",
  "beautifulsoup4>=4.12",
  "PyYAML>=6.0",
  "Jinja2>=3.0",
  "reportlab>=4.0",
  "pyjwt>=2.9,<2.10",
  "cryptography>=42",
  "python-docx>=1.1",
  "ftfy>=6.0",
  "pikepdf>=8.0"
]

[project.scripts]
annex4ac = "annex4ac:app"

[build-system]                # PEP 517
requires = ["setuptools>=68", "wheel", "build"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = "-q"

[tool.setuptools.packages.find]
include = ["annex4ac"]

[tool.setuptools.package-data]
annex4ac = ["resources/*.icc", "fonts/*.ttf", "templates/*.html"]